requirements.txt
absl-py==2.0.0
astunparse==1.6.3
beautifulsoup4==4.12.2
breadability==0.1.20
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
datefinder==0.7.3
docopt==0.6.2
et-xmlfile==1.1.0
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.59.3
h5py==3.10.0
idna==3.6
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
lxml==5.0.0
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==23.12.11
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.24
summa==1.2.0
sumy==0.11.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-cpu
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
